requirements.txt
discord.py
requests
flask
google-genai
python-dotenv
